feat: add two permission classes
CreateUserOrIsColaborator',  override
'has_permission' method from 'BasePermission' and verify if request method is 'POST'. If not, the permission is only granted if the user is a superuser. 'IsOwnerOrColaborator, override 'has_permission' method from 'BasePermission' and check if the token id is equal to id sent in url. If not, the permission is only granted if the user is a superuser